feat : move to component, set draggable false
